+++ medusa/tv/series.py
@@ -3,7 +3,6 @@
 """Series classes."""
 from __future__ import unicode_literals
 
-import ast
 import copy
 import datetime
 import glob
@@ -29,6 +28,7 @@ from medusa import (
     network_timezones,
     notifiers,
     post_processor,
+    ui,
 )
 from medusa.black_and_white_list import BlackAndWhiteList
 from medusa.common import (
@@ -49,11 +49,7 @@ from medusa.common import (
     qualityPresets,
     statusStrings,
 )
-from medusa.helper.common import (
-    episode_num,
-    pretty_file_size,
-    try_int,
-)
+from medusa.helper.common import episode_num, pretty_file_size, sanitize_filename, try_int
 from medusa.helper.exceptions import (
     CantRefreshShowException,
     CantRemoveShowException,
@@ -64,8 +60,9 @@ from medusa.helper.exceptions import (
     ShowNotFoundException,
     ex,
 )
+from medusa.helpers import make_dir
 from medusa.helpers.anidb import short_group_names
-from medusa.helpers.externals import get_externals, load_externals_from_db
+from medusa.helpers.externals import check_existing_shows, get_externals, load_externals_from_db
 from medusa.helpers.utils import dict_to_array, safe_get, to_camel_case
 from medusa.imdb import Imdb
 from medusa.indexers.api import indexerApi
@@ -75,9 +72,7 @@ from medusa.indexers.config import (
     indexerConfig
 )
 from medusa.indexers.exceptions import (
-    IndexerAttributeNotFound,
-    IndexerException,
-    IndexerSeasonNotFound,
+    IndexerAttributeNotFound, IndexerException, IndexerSeasonNotFound, IndexerShowAlreadyInLibrary
 )
 from medusa.indexers.tmdb.api import Tmdb
 from medusa.indexers.utils import (
@@ -102,7 +97,7 @@ from medusa.scene_exceptions import get_all_scene_exceptions, get_scene_exceptio
 from medusa.scene_numbering import (
     get_scene_absolute_numbering_for_show, get_scene_numbering_for_show,
     get_xem_absolute_numbering_for_show, get_xem_numbering_for_show,
-    numbering_tuple_to_dict,
+    numbering_tuple_to_dict, xem_refresh
 )
 from medusa.search import FORCED_SEARCH
 from medusa.show.show import Show
@@ -114,7 +109,7 @@ from medusa.tv.base import Identifier, TV
 from medusa.tv.episode import Episode
 from medusa.tv.indexer import Indexer
 
-from six import iteritems, itervalues, string_types, text_type, viewitems
+from six import ensure_text, iteritems, itervalues, string_types, text_type, viewitems
 
 import ttl_cache
 
@@ -128,6 +123,10 @@ MILLIS_YEAR_1900 = datetime.datetime(year=1900, month=1, day=1).toordinal()
 
 log = CustomBraceAdapter(logging.getLogger(__name__))
 log.logger.addHandler(logging.NullHandler())
+
+
+class SaveSeriesException(Exception):
+    """Generic exception used for adding a new series."""
 
 
 class SeriesIdentifier(Identifier):
@@ -163,10 +162,19 @@ class SeriesIdentifier(Identifier):
         """Slug."""
         return text_type(self)
 
-    @property
-    def api(self):
-        """Api."""
+    def get_indexer_api(self, options=None):
+        """Return an indexer api for this show."""
         indexer_api = indexerApi(self.indexer.id)
+        indexer_api_params = indexer_api.api_params.copy()
+
+        if options and options.get('language') is not None:
+            indexer_api_params['language'] = options['language']
+
+        log.debug('{indexer_name}: {indexer_params!r}', {
+            'indexer_name': indexerApi(self.indexer.id).name,
+            'indexer_params': indexer_api_params
+        })
+
         return indexer_api.indexer(**indexer_api.api_params)
 
     def __bool__(self):
@@ -278,25 +286,9 @@ class Series(TV):
         """Create a series object from its identifier."""
         return Series(identifier.indexer.id, identifier.id)
 
-    # TODO: Make this the single entry to add new series
-    @classmethod
-    def save_series(cls, series):
-        """Save the specified series to medusa."""
-        try:
-            api = series.identifier.api
-            series.load_from_indexer(tvapi=api)
-            series.load_imdb_info()
-            app.showList.append(series)
-            series.save_to_db()
-            series.load_episodes_from_indexer(tvapi=api)
-            series.populate_cache()
-            return series
-        except IndexerException as error:
-            log.warning('Unable to load series from indexer: {0!r}'.format(error))
-
     @property
     def identifier(self):
-        """Identifier."""
+        """Return a series identifier object."""
         return SeriesIdentifier(self.indexer, self.series_id)
 
     @property
@@ -1482,7 +1474,7 @@ class Series(TV):
             self.sports = int(sql_results[0]['sports'] or 0)
             self.scene = int(sql_results[0]['scene'] or 0)
             self.subtitles = int(sql_results[0]['subtitles'] or 0)
-            self.notify_list = dict(ast.literal_eval(sql_results[0]['notify_list'] or '{}'))
+            self.notify_list = json.loads(sql_results[0]['notify_list'] or '{}')
             self.dvd_order = int(sql_results[0]['dvdorder'] or 0)
             self.quality = int(sql_results[0]['quality'] or Quality.NA)
             self.season_folders = int(not (sql_results[0]['flatten_folders'] or 0))  # TODO: Rename this in the DB
@@ -1659,6 +1651,95 @@ class Series(TV):
         log.debug(u'{id}: Obtained info from IMDb: {imdb_info}',
                   {'id': self.series_id, 'imdb_info': self.imdb_info})
 
+    def check_existing(self):
+        """Check if we can already find this show in our current showList."""
+        try:
+            check_existing_shows(self.identifier.get_indexer_api(), self.identifier.indexer.id)
+        except IndexerShowAlreadyInLibrary as error:
+            log.warning(
+                'Could not add the show {series}, as it already is in your library.'
+                'Error: {error}',
+                {'series': self.name, 'error': error}
+            )
+            ui.notifications.error(
+                'Unable to add show',
+                'reason: {0}'.format(error)
+            )
+            raise SaveSeriesException(
+                'Unable to add show',
+                'reason: {0}'.format(error)
+            )
+
+    def configure(self, queue_item):
+        """Configure series."""
+        # Let's try to create the show Dir if it's not provided. This way we force the show dir
+        # to build using the Indexers provided series name
+        options = queue_item.options
+        show_dir = queue_item.show_dir
+        root_dir = queue_item.root_dir or app.ROOT_DIRS[int(app.ROOT_DIRS[0]) + 1] if len(app.ROOT_DIRS) > 0 else None
+
+        if not show_dir and root_dir:
+            # I don't think we need to check this. We just create the folder after we already queried the indexer.
+            # show_name = get_showname_from_indexer(self.indexer, self.indexer_id, self.lang)
+            if self.name:
+                show_dir = os.path.join(root_dir, sanitize_filename(self.name))
+
+                # don't create show dir if config says not to
+                if app.ADD_SHOWS_WO_DIR:
+                    log.info('Skipping initial creation of {path} due to config.ini setting',
+                             {'path': show_dir})
+                else:
+                    dir_exists = make_dir(show_dir)
+                    if not dir_exists:
+                        log.error("Unable to create the folder {0}, can't add the show", show_dir)
+                        ui.notifications.error(
+                            'Unable to create folder',
+                            'folder: {0}'.format(show_dir)
+                        )
+                        return
+
+                    helpers.chmod_as_parent(show_dir)
+            else:
+                log.error("Unable to get a show {0}, can't add the show", show_dir)
+                return
+
+        if show_dir:
+            self.location = ensure_text(show_dir)
+
+        self.subtitles = options['subtitles'] if options.get('subtitles') is not None else app.SUBTITLES_DEFAULT
+
+        if options.get('quality'):
+            self.qualities_allowed = options['quality']['allowed']
+            self.qualities_preferred = options['quality']['preferred']
+        else:
+            self.qualities_allowed, self.qualities_preferred = Quality.split_quality(int(app.QUALITY_DEFAULT))
+
+        self.season_folders = options['season_folders'] if options.get('season_folders') is not None \
+            else app.SEASON_FOLDERS_DEFAULT
+        self.anime = options['anime'] if options.get('anime') is not None else app.ANIME_DEFAULT
+        self.scene = options['scene'] if options.get('scene') is not None else app.SCENE_DEFAULT
+        self.paused = options['paused'] if options.get('paused') is not None else False
+        self.lang = options['language'] if options.get('language') is not None else app.INDEXER_DEFAULT_LANGUAGE
+        self.show_lists = options['show_lists'] if options.get('show_lists') is not None else app.SHOWLISTS_DEFAULT
+
+        if options.get('default_status') is not None:
+            # set up default new/missing episode status
+            log.info(
+                'Setting all previously aired episodes to the specified status: {status}',
+                {'status': statusStrings[options['default_status']]}
+            )
+            self.default_ep_status = options['default_status']
+        else:
+            self.default_ep_status = app.STATUS_DEFAULT
+
+        if self.anime:
+            self.release_groups = BlackAndWhiteList(self)
+            if options.get('release') is not None:
+                if options['release']['blacklist']:
+                    self.release_groups.set_black_keywords(options['release']['blacklist'])
+                if options['release']['whitelist']:
+                    self.release_groups.set_white_keywords(options['release']['whitelist'])
+
     def prev_episode(self):
         """Return the last aired episode air date.
 
@@ -1828,6 +1909,43 @@ class Series(TV):
         log.debug(u'{id}: Checking & filling cache for show {show}',
                   {'id': self.series_id, 'show': self.name})
         image_cache.fill_cache(self)
+
+    def sync_trakt(self):
+        """Sync trakt episodes if trakt is enabled."""
+        if app.USE_TRAKT:
+            # if there are specific episodes that need to be added by trakt
+            app.trakt_checker_scheduler.action.manage_new_show(self)
+
+            # add show to trakt.tv library
+            if app.TRAKT_SYNC:
+                app.trakt_checker_scheduler.action.add_show_trakt_library(self)
+
+            if app.TRAKT_SYNC_WATCHLIST:
+                log.info('updating trakt watchlist')
+                notifiers.trakt_notifier.update_watchlist(show_obj=self)
+
+    def add_scene_numbering(self):
+        """
+        Add XEM data to DB for show.
+
+        Warn the user if an episode number mapping is available on thexem.
+        Only use this method, through show_queue.QueueItemAdd().
+        """
+        xem_refresh(self, force=True)
+
+        # check if show has XEM mapping so we can determine if searches
+        # should go by scene numbering or indexer numbering. Warn the user.
+        if not self.scene and get_xem_numbering_for_show(self):
+            log.warning(
+                '{id}: while adding the show {title} we noticed thexem.de has an episode mapping available'
+                '\nyou might want to consider enabling the scene option for this show.',
+                {'id': self.series_id, 'title': self.name}
+            )
+            ui.notifications.message(
+                'consider enabling scene for this show',
+                'for show {title} you might want to consider enabling the scene option'
+                .format(title=self.name)
+            )
 
     def refresh_dir(self):
         """Refresh show using its location.
@@ -2261,7 +2379,8 @@ class Series(TV):
         """
         Configure the show lists, for this show.
 
-        :param show_lists: Comma separated list of show categories.
+        self._show_lists is stored as a comma separated string.
+        :param show_lists: A list of show categories.
         """
         self._show_lists = ','.join(show_lists) if show_lists else 'series'