+++ themes-default/slim/src/store/modules/config/general.js
@@ -37,7 +37,8 @@ const state = {
         subtitles: null,
         seasonFolders: null,
         anime: null,
-        scene: null
+        scene: null,
+        showLists: null
     },
     launchBrowser: null,
     defaultPage: null,